Added tests for the CUDA case and refactored the testing suite to eliminate code duplication
